Do not track the "warning" state
"Warning" is not a true state we should track, since it does not trigger any events.
Its purpose is simply to prevent spurious error alerts due to jitter in ping arrival times.
If we were to record "Warning" states, we would end up with a lot of noise when "Warning" is entered for a few seconds at a time.
"Warning" should be treated as a sub-state of "Ok" that is merely distinguished visually in the dashboard.
Instead of "Warning", it should probably be called "Late".